Include aws-cfn-resource-specs in docs
Adding a link to a repository that gives expanded details from audits of the latest AWS CloudFormation specification files. This repository works as a more detailed version of the Release History page provided in the AWS CFN User Guide, including:
- Broken documentation links
- Missing user guide documentation in the GitHub repo mirror
- Both human readable and machine readable changelog information about each specification release. Tracking:
  - Introductions of new resource and property types
  - Existing types added as supported in new regions
  - Tracking of when types are deleted/removed (this could also mean _renamed_)
  - Tracking of types missing from `us-east-1`, if any

This is helpful for anyone wanting to have an easy view of what's changed in each version release, mostly via the `CHANGELOG.md` file. It will likely be most useful to people who manage code generation tools and SDKs wrapped around the ingestion of the specification files.