Update README.md (#11)
* Update README.md

hi, i have a tool that i developed, but am also using on a number of sizeable AWS Organizations. I really enjoy working on this and am looking forward to have more people using it. would be really happy to get included in the list.

thanks!

* Update README.md

fixed casing of CloudFormation